requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
distro==1.0.3
Django==1.10.6
djangorestframework==3.6.2
dlib==19.4.0
face-recognition==0.1.10
face-recognition-models==0.1.3
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
numpy==1.12.1
oauthlib==1.0.3
olefile==0.44
onboard==1.2.0
opencv-python==3.2.0.6
padme==1.1.1
pexpect==4.0.1
Pillow==4.0.0
plainbox==0.25
psycopg2==2.7.1
ptyprocess==0.5
pyasn1==0.1.9
PyAudio==0.2.8
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
scipy==0.19.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
SpeechRecognition==3.6.3
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
